# public/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Plan-B API",
    "version": "0.2.0",
    "description": "Grasscutter Web Management Platform"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "GiveItem": {
        "type": "object",
        "required": [
          "itemId"
        ],
        "properties": {
          "itemId": {
            "type": "integer"
          },
          "amount": {
            "type": "integer",
            "default": 1
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/players": {
      "get": {
        "tags": [
          "Players"
        ],
        "summary": "Player list",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Player array"
          }
        }
      }
    },
    "/api/players/{uid}": {
      "get": {
        "tags": [
          "Players"
        ],
        "summary": "Player detail",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full player data"
          }
        }
      }
    },
    "/api/players/{uid}/give": {
      "post": {
        "tags": [
          "Players"
        ],
        "summary": "Give items via GM",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GiveItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result"
          }
        }
      }
    },
    "/api/players/{uid}/ban": {
      "post": {
        "tags": [
          "Players"
        ],
        "summary": "Ban player",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/players/{uid}/unban": {
      "post": {
        "tags": [
          "Players"
        ],
        "summary": "Unban player",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/players/{uid}/kick": {
      "post": {
        "tags": [
          "Players"
        ],
        "summary": "Kick player",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/players/{uid}/teleport": {
      "post": {
        "tags": [
          "Players"
        ],
        "summary": "Teleport player",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/players/{uid}/setlevel": {
      "post": {
        "tags": [
          "Players"
        ],
        "summary": "Set adventure rank",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/players/{uid}/setworldlevel": {
      "post": {
        "tags": [
          "Players"
        ],
        "summary": "Set world level",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/gacha": {
      "get": {
        "tags": [
          "Gacha"
        ],
        "summary": "All banners",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Banner array"
          }
        }
      },
      "post": {
        "tags": [
          "Gacha"
        ],
        "summary": "Save banners",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/gacha/items/search": {
      "get": {
        "tags": [
          "Gacha"
        ],
        "summary": "Search gacha items (mappings.js)",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rarity",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Items"
          }
        }
      }
    },
...
    "/api/multiplayer/players": {
      "get": {
        "tags": [
          "Multiplayer"
        ],
        "summary": "Player locations",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Players + positions"
          }
        }
      }
    },
...
    "/api/teleport/player": {
      "post": {
        "tags": [
          "System"
        ],
        "summary": "Teleport player to preset",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}